feat(frontend): use new Solana transaction service
