+++ src/frontend/src/sol/api/solana.api.ts
@@ -1,15 +1,12 @@
-import { WALLET_PAGINATION } from '$lib/constants/app.constants';
 import type { OptionSolAddress, SolAddress } from '$lib/types/address';
 import { last } from '$lib/utils/array.utils';
 import { ATA_SIZE } from '$sol/constants/ata.constants';
 import { solanaHttpRpc } from '$sol/providers/sol-rpc.providers';
 import type { SolanaNetworkType } from '$sol/types/network';
-import type { GetSolTransactionsParams } from '$sol/types/sol-api';
-import type { SolRpcTransaction, SolSignature } from '$sol/types/sol-transaction';
-import { getSplBalanceChange } from '$sol/utils/spl-transactions.utils';
+import type { SolSignature } from '$sol/types/sol-transaction';
 import { isNullish, nonNullish } from '@dfinity/utils';
-import { address, assertIsAddress, address as solAddress, type Address } from '@solana/addresses';
-import { signature, type Signature } from '@solana/keys';
+import { address, address as solAddress, type Address } from '@solana/addresses';
+import { type Signature } from '@solana/keys';
 import type { Lamports } from '@solana/rpc-types';
 import type { Writeable } from 'zod';
 
@@ -168,75 +165,6 @@ export const loadTokenAccount = async ({
 	const { pubkey: accountAddress } = response.value[0];
 
 	return accountAddress;
-};
-
-/**
- * Fetches SPL token transactions for a given wallet address and token mint.
- */
-//TODO add unit tests
-export const getSplTransactions = async ({
-	address,
-	network,
-	tokenAddress,
-	before,
-	limit = Number(WALLET_PAGINATION)
-}: GetSolTransactionsParams & {
-	tokenAddress: SolAddress;
-}): Promise<SolRpcTransaction[]> => {
-	assertIsAddress(tokenAddress);
-
-	const { getTokenAccountsByOwner } = solanaHttpRpc(network);
-	const relevantTokenAddress = solAddress(tokenAddress);
-	const wallet = solAddress(address);
-
-	const beforeSignature = nonNullish(before) ? signature(before) : undefined;
-
-	const tokenAccounts = await getTokenAccountsByOwner(
-		wallet,
-		{
-			mint: relevantTokenAddress
-		},
-		{ encoding: 'jsonParsed' }
-	).send();
-
-	const signatures = (
-		await Promise.all(
-			tokenAccounts.value.map(({ pubkey }) =>
-				fetchSignatures({
-					network,
-					wallet: solAddress(pubkey),
-					before: beforeSignature,
-					limit
-				})
-			)
-		)
-	).flat();
-
-	const transactions = await signatures.reduce(
-		async (accPromise, sig) => {
-			const acc = await accPromise;
-			const transactionDetail = await fetchTransactionDetailForSignature({
-				signature: sig,
-				network
-			});
-
-			if (
-				nonNullish(transactionDetail) &&
-				//TODO handle self sending
-				getSplBalanceChange({
-					transaction: transactionDetail,
-					tokenAddress,
-					address
-				}) > 0
-			) {
-				acc.push(transactionDetail);
-			}
-			return acc;
-		},
-		Promise.resolve([] as SolRpcTransaction[])
-	);
-
-	return transactions.slice(0, limit);
 };
 
 /**
