+++ src/frontend/src/sol/schedulers/sol-wallet.scheduler.ts
@@ -7,18 +7,12 @@ import type {
 } from '$lib/types/post-message';
 import type { CertifiedData } from '$lib/types/store';
 import type { Option } from '$lib/types/utils';
-import {
-	getSplTransactions,
-	loadSolLamportsBalance,
-	loadSplTokenBalance
-} from '$sol/api/solana.api';
-import { getSolTransactions } from '$sol/services/sol-signatures.services';
+import { loadSolLamportsBalance, loadSplTokenBalance } from '$sol/api/solana.api';
+import { getSolTransactions, getSplTransactions } from '$sol/services/sol-signatures.services';
 import type { SolCertifiedTransaction } from '$sol/stores/sol-transactions.store';
 import type { SolanaNetworkType } from '$sol/types/network';
 import type { SolBalance } from '$sol/types/sol-balance';
 import type { SolPostMessageDataResponseWallet } from '$sol/types/sol-post-message';
-import { mapSolTransactionUi } from '$sol/utils/sol-transactions.utils';
-import { mapSplTransactionUi } from '$sol/utils/spl-transactions.utils';
 import { assertNonNullish, isNullish, jsonReplacer, nonNullish } from '@dfinity/utils';
 
 interface LoadSolWalletParams {
@@ -90,13 +84,7 @@ export class SolWalletScheduler implements Scheduler<PostMessageDataRequestSol> 
 			: await getSolTransactions({ network: solanaNetwork, address });
 
 		const transactionsUi = transactions.map((transaction) => ({
-			data: nonNullish(tokenAddress)
-				? mapSplTransactionUi({
-						transaction,
-						tokenAddress,
-						address
-					})
-				: mapSolTransactionUi({ transaction, address }),
+			data: transaction,
 			certified: false
 		}));
