+++ src/frontend/src/sol/services/sol-signatures.services.ts
@@ -1,10 +1,11 @@
 import { WALLET_PAGINATION } from '$lib/constants/app.constants';
-import { fetchSignatures, fetchTransactionDetailForSignature } from '$sol/api/solana.api';
+import type { SolAddress } from '$lib/types/address';
+import { fetchSignatures } from '$sol/api/solana.api';
+import { fetchSolTransactionsForSignature } from '$sol/services/sol-transactions.services';
 import type { GetSolTransactionsParams } from '$sol/types/sol-api';
-import type { SolRpcTransaction } from '$sol/types/sol-transaction';
-import { getSolBalanceChange } from '$sol/utils/sol-transactions.utils';
+import type { SolTransactionUi } from '$sol/types/sol-transaction';
 import { nonNullish } from '@dfinity/utils';
-import { address as solAddress } from '@solana/addresses';
+import { assertIsAddress, address as solAddress } from '@solana/addresses';
 import { signature } from '@solana/keys';
 
 /**
@@ -15,7 +16,7 @@ export const getSolTransactions = async ({
 	network,
 	before,
 	limit = Number(WALLET_PAGINATION)
-}: GetSolTransactionsParams): Promise<SolRpcTransaction[]> => {
+}: GetSolTransactionsParams): Promise<SolTransactionUi[]> => {
 	const wallet = solAddress(address);
 	const beforeSignature = nonNullish(before) ? signature(before) : undefined;
 	const signatures = await fetchSignatures({ network, wallet, before: beforeSignature, limit });
@@ -23,16 +24,54 @@ export const getSolTransactions = async ({
 	const transactions = await signatures.reduce(
 		async (accPromise, signature) => {
 			const acc = await accPromise;
-			const transactionDetail = await fetchTransactionDetailForSignature({ signature, network });
-			if (
-				nonNullish(transactionDetail) &&
-				getSolBalanceChange({ transaction: transactionDetail, address })
-			) {
-				acc.push(transactionDetail);
-			}
-			return acc;
+			const parsedTransactions = await fetchSolTransactionsForSignature({
+				signature,
+				network,
+				address
+			});
+
+			return [...acc, ...parsedTransactions];
 		},
-		Promise.resolve([] as SolRpcTransaction[])
+		Promise.resolve([] as SolTransactionUi[])
+	);
+
+	return transactions.slice(0, limit);
+};
+
+/**
+ * Fetches SPL token transactions for a given wallet address and token mint.
+ */
+//TODO add unit tests
+export const getSplTransactions = async ({
+	address,
+	network,
+	tokenAddress,
+	before,
+	limit = Number(WALLET_PAGINATION)
+}: GetSolTransactionsParams & {
+	tokenAddress: SolAddress;
+}): Promise<SolTransactionUi[]> => {
+	assertIsAddress(tokenAddress);
+
+	const wallet = solAddress(address);
+
+	const beforeSignature = nonNullish(before) ? signature(before) : undefined;
+
+	const signatures = await fetchSignatures({ network, wallet, before: beforeSignature, limit });
+
+	const transactions = await signatures.reduce(
+		async (accPromise, signature) => {
+			const acc = await accPromise;
+			const parsedTransactions = await fetchSolTransactionsForSignature({
+				signature,
+				network,
+				address,
+				tokenAddress
+			});
+
+			return [...acc, ...parsedTransactions];
+		},
+		Promise.resolve([] as SolTransactionUi[])
 	);
 
 	return transactions.slice(0, limit);
